change class names
